+++ R/bigstatsr-package.R
@@ -3,7 +3,6 @@
 #' @useDynLib bigstatsr, .registration = TRUE
 #' @importFrom Rcpp sourceCpp
 #' @importFrom methods new
-#' @importFrom bigparallelr makeCluster stopCluster registerDoParallel
 #' @import foreach
 #'
 #' @param X A [FBM][FBM-class].